+++ src/main/java/net/csibio/aird/bean/SpectrumDetail.java
@@ -6,7 +6,7 @@ import lombok.Data;
 public class SpectrumDetail {
 
   float rt;
-  float[] mzs;
+  double[] mzs;
   float[] intensities;
   byte[] mzBytes;    //压缩前的数mz据流
   byte[] intensityBytes;   //压缩前的数intensity据流
